Move AblyEventMessage into a separate file to avoid errors in Android Studio linting
