+++ android/src/main/java/io/ably/flutter/plugin/AblyFlutterMessage.java
@@ -21,18 +21,3 @@ public class AblyFlutterMessage<T> {
     }
 
 }
-
-
-class AblyEventMessage<T> {
-    final String eventName;
-    final T message;
-
-    AblyEventMessage(final String eventName, final T message) {
-        if (null == eventName) {
-            throw new NullPointerException("message cannot be null.");
-        }
-        this.eventName = eventName;
-        this.message = message;
-    }
-
-}
